update `GeoClientEnCours` entity #18056

+++ src/main/java/fr/microtec/geo2/persistance/entity/tiers/GeoClientEnCours.java
@@ -17,7 +17,7 @@ public class GeoClientEnCours {
 
     @Id
     @Column(name = "rownum")
-    private Float id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "soc_code")
